Replace user-based contacts with account-based
This changes the URL used from /v1.1/contacts (contacts on the user) to
/v2.0/contacts (contacts on the account).

The fields on Contact have been updated to reflect the new resource.
Calls to get all contacts or to create a contact now require an account
reference.

The method to create multiple contacts has been removed as this feature
has been removed from the REST API.